Replaced "double" with float" almost everywhere. Exceptions include:
-front end was not touched.
-numbers in linear scale that get converted to log, and numbers in
 log that get converted to linear. Essentially, "linear" is double and
 "log" is float.


git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@1473 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ edu/cmu/sphinx/decoder/linguist/FSTGrammar.java
@@ -221,7 +221,7 @@ public class FSTGrammar extends Grammar {
 		    assert hasWord(nextNode);
 		}
 		
-		thisNode.add(nextNode, (float) convertProbability(lnProb));
+		thisNode.add(nextNode, convertProbability(lnProb));
 		
 	    } else if (token.equals("F")) {
 		int thisID = tok.getInt("this id");
@@ -234,7 +234,7 @@ public class FSTGrammar extends Grammar {
 		    thisNode = getEndNode(thisNode);
 		}
 		
-		thisNode.add(nextNode, (float) convertProbability(lnProb));
+		thisNode.add(nextNode, convertProbability(lnProb));
 	    }
 	}
 	tok.close();
@@ -351,7 +351,7 @@ public class FSTGrammar extends Grammar {
      * 
      * @return the converted probability in logMath log base
      */
-    private double convertProbability(double lnProb) {
+    private float convertProbability(float lnProb) {
         return getLogMath().lnToLog(-lnProb);
     }
